// src/interfaces/Filter.ts
export interface IFilter {
    areaIds: number[] | null;
    rooms: number[] | null;
    complexIds: number[] | null;
    entrance: number | null
    minFloor: number | null;
    maxFloor: number | null;
    minPrice: number | null;
    maxPrice: number | null;
    minArea: number | null;
    maxArea: number | null;
  }